refactor(entry-points): drop legacy ml_playground.data shim
