+++ ml_playground/training/hooks/data.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from pathlib import Path
 
 from ml_playground.configuration import TrainerConfig, SharedConfig
-from ml_playground.data import SimpleBatches
+from ml_playground.data_pipeline.sampling.batches import SimpleBatches
 
 
 __all__ = ["initialize_batches"]
