+++ ml_playground/training/hooks/evaluation.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from torch.utils.tensorboard import SummaryWriter
 
 from ml_playground.configuration import TrainerConfig
-from ml_playground.data import SimpleBatches
+from ml_playground.data_pipeline.sampling.batches import SimpleBatches
 from ml_playground.estimator import estimate_loss
 from ml_playground.models.core.model import GPT
